requirements.txt
revChatGPT==1.2.1
tls_client
undetected_chromedriver
replicate
python-dotenv